agents/intake_agent/pyproject.toml
[tool.poetry]
name = "intake-agent"
version = "0.1.0"
description = "intake_agent microservice"
authors = ["Agentic Platform Team"]

[[tool.poetry.packages]]
include = "src"

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
langgraph = "^0.0.40"
pydantic = "^2.6.0"
sqlacodegen = "^3.2.0"
psycopg2-binary = "^2.9.11"
psycopg2 = "^2.9.11"
sqlalchemy = "^2.0.46"
pydantic-settings = "^2.12.0"
asyncpg = "^0.31.0"
alembic = "^1.18.3"
ua-parser = "^0.18.0"
numpy = "^2.4.2"
opencv-python = "^4.13.0.90"
python-multipart = "^0.0.22"
pillow = "^12.1.0"
transformers = "^5.0.0"
tokenizers = "^0.22.2"
torch = "^2.10.0"
pytesseract = "^0.3.13"
pdf2image = "^1.17.0"

[tool.poetry.scripts]
dev = "src.cli:dev"
prod = "src.cli:prod"
test = "src.cli:test"
migration = "src.cli:migration"
initial_migration = "src.cli:migrate"
migrate_new_model = "src.cli:migrate_new_model"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.3.0"
pytest-asyncio = "^0.23"
python-dotenv = "^1.2.1"


[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]
python_files = "test_*.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"